# Downgrade to z3 4.12.6 and adjust CBMC proofs