--- a/mldsa/polyvec.c
+++ b/mldsa/polyvec.c
@@ -489,8 +489,9 @@ void mld_polyveck_decompose(mld_polyveck *v1, mld_polyveck *v0,
     assigns(i, memory_slice(v0, sizeof(mld_polyveck)), memory_slice(v1, sizeof(mld_polyveck)))
     invariant(i <= MLDSA_K)
     invariant(forall(k1, 0, i,
-                     array_bound(v1->vec[k1].coeffs, 0, MLDSA_N, 0, (MLDSA_Q-1)/(2*MLDSA_GAMMA2)) &&
-                     array_abs_bound(v0->vec[k1].coeffs, 0, MLDSA_N, MLDSA_GAMMA2+1)))
+                     array_bound(v1->vec[k1].coeffs, 0, MLDSA_N, 0, (MLDSA_Q-1)/(2*MLDSA_GAMMA2))))
+    invariant(forall(k2, 0, i,
+                     array_abs_bound(v0->vec[k2].coeffs, 0, MLDSA_N, MLDSA_GAMMA2+1)))
   )
   {
     mld_poly_decompose(&v1->vec[i], &v0->vec[i], &v->vec[i]);
@@ -539,6 +540,10 @@ void mld_polyveck_pack_w1(uint8_t r[MLDSA_K * MLDSA_POLYW1_PACKEDBYTES],
   unsigned int i;
 
   for (i = 0; i < MLDSA_K; ++i)
+  __loop__(
+    assigns(i, object_whole(r))
+    invariant(i <= MLDSA_K)
+  )
   {
     mld_polyw1_pack(&r[i * MLDSA_POLYW1_PACKEDBYTES], &w1->vec[i]);
   }
@@ -549,8 +554,12 @@ void mld_polyveck_pack_eta(uint8_t r[MLDSA_K * MLDSA_POLYETA_PACKEDBYTES],
 {
   unsigned int i;
   for (i = 0; i < MLDSA_K; ++i)
+  __loop__(
+    assigns(i, object_whole(r))
+    invariant(i <= MLDSA_K)
+  )
   {
-    mld_polyeta_pack(r + i * MLDSA_POLYETA_PACKEDBYTES, &p->vec[i]);
+    mld_polyeta_pack(&r[i * MLDSA_POLYETA_PACKEDBYTES], &p->vec[i]);
   }
 }
 
@@ -559,8 +568,12 @@ void mld_polyvecl_pack_eta(uint8_t r[MLDSA_L * MLDSA_POLYETA_PACKEDBYTES],
 {
   unsigned int i;
   for (i = 0; i < MLDSA_L; ++i)
+  __loop__(
+    assigns(i, object_whole(r))
+    invariant(i <= MLDSA_L)
+  )
   {
-    mld_polyeta_pack(r + i * MLDSA_POLYETA_PACKEDBYTES, &p->vec[i]);
+    mld_polyeta_pack(&r[i * MLDSA_POLYETA_PACKEDBYTES], &p->vec[i]);
   }
 }
 
@@ -569,8 +582,12 @@ void mld_polyvecl_pack_z(uint8_t r[MLDSA_L * MLDSA_POLYZ_PACKEDBYTES],
 {
   unsigned int i;
   for (i = 0; i < MLDSA_L; ++i)
+  __loop__(
+    assigns(i, object_whole(r))
+    invariant(i <= MLDSA_L)
+  )
   {
-    mld_polyz_pack(r + i * MLDSA_POLYZ_PACKEDBYTES, &p->vec[i]);
+    mld_polyz_pack(&r[i * MLDSA_POLYZ_PACKEDBYTES], &p->vec[i]);
   }
 }
 
@@ -580,8 +597,12 @@ void mld_polyveck_pack_t0(uint8_t r[MLDSA_K * MLDSA_POLYT0_PACKEDBYTES],
 {
   unsigned int i;
   for (i = 0; i < MLDSA_K; ++i)
+  __loop__(
+    assigns(i, object_whole(r))
+    invariant(i <= MLDSA_K)
+  )
   {
-    mld_polyt0_pack(r + i * MLDSA_POLYT0_PACKEDBYTES, &p->vec[i]);
+    mld_polyt0_pack(&r[i * MLDSA_POLYT0_PACKEDBYTES], &p->vec[i]);
   }
 }