--- a/proofs/cbmc/polyveck_decompose/Makefile
+++ b/proofs/cbmc/polyveck_decompose/Makefile
@@ -27,6 +27,7 @@ USE_DYNAMIC_FRAMES=1
 # Disable any setting of EXTERNAL_SAT_SOLVER, and choose SMT backend instead
 EXTERNAL_SAT_SOLVER=
 CBMCFLAGS=--smt2
+CBMCFLAGS += --slice-formula
 
 FUNCTION_NAME = polyveck_decompose