We have noticed a severe performance degradation of some CBMC proofs when moving from z3 4.12 to 4.13/4.14/4.15. This is currently blocking the proof of the signing functions.
This PR downgrades back to z3 4.12.6 - the last release prior to the performance regression. A few minor changes to the other CBMC proofs are required which are applied as part of this PR.