@mkannwischer 2025-07-17
Changes in proof performance:
- ML-DSA-44: 12 minutes -> 21 minutes
- ML-DSA-65: 12 minutes -> 8 minutes
- ML-DSA-87: 13 minutes -> 8 minutes

Major bottleneck appears to be `polyvecl_pointwise_acc_montgomery` proof
 - ML-DSA-44: 268s -> 1068s
 - ML-DSA-65: 230s -> 269s
 - ML-DSA-87: 307s -> 238s
 
 Also `poly_pointwise_montgomery`:
 - ML-DSA-44: 71s -> 285s
 - ML-DSA-65: 56s -> 72s
 - ML-DSA-87: 104s -> 99s